+++ verticapy/sdk/vertica/udf/__init__.py
@@ -14,4 +14,4 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.sdk.vertica.udf import import_lib_udf, generate_lib_udf
+from verticapy.sdk.vertica.udf.gen import import_lib_udf, generate_lib_udf, generate_udf
